Fix typing bugs in io.h read_port functions

+++ include/arch/x86/io.h
@@ -14,10 +14,10 @@ void write_port32(uint32_t value, uint16_t port);
 /* Read from an I/O port.
  * A wrapper around inb. Takes a port and returns its value.
  */
-uint16_t read_port(uint8_t port);
+uint16_t read_port(uint16_t port);
 /* Read two bytes from an I/O port. */
 uint16_t read_port16(uint16_t port);
 /* Read a word from an I/O port. */
-uint16_t read_port32(uint32_t port);
+uint32_t read_port32(uint16_t port);
 
 #endif
